requirements.txt
scipy # As mentioned in the last error.
Pillow==10.0.0
numpy
opencv-python
torch
torchvision
mlflow==1.20.2


